// lib/api/schemas/companyResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { CompanyResponseIndustry } from './companyResponseIndustry';
import type { CompanyResponseSize } from './companyResponseSize';
import type { CompanyResponseWebsite } from './companyResponseWebsite';
import type { CompanyResponseDescription } from './companyResponseDescription';
import type { CompanyResponsePrimaryColor } from './companyResponsePrimaryColor';
import type { CompanyResponseSecondaryColor } from './companyResponseSecondaryColor';
import type { CompanyResponseAccentColor } from './companyResponseAccentColor';
import type { CompanyResponseLogoUrl } from './companyResponseLogoUrl';

export interface CompanyResponse {
  name: string;
  industry?: CompanyResponseIndustry;
  size?: CompanyResponseSize;
  website?: CompanyResponseWebsite;
  description?: CompanyResponseDescription;
  primary_color?: CompanyResponsePrimaryColor;
  secondary_color?: CompanyResponseSecondaryColor;
  accent_color?: CompanyResponseAccentColor;
  id: string;
  logo_url?: CompanyResponseLogoUrl;
  created_at: string;
  updated_at: string;
}