// lib/api/schemas/companyUpdate.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Onboarding-as-a-Service
 * Onboarding-as-a-Service (OaaS) Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { CompanyUpdateName } from './companyUpdateName';
import type { CompanyUpdateIndustry } from './companyUpdateIndustry';
import type { CompanyUpdateSize } from './companyUpdateSize';
import type { CompanyUpdateWebsite } from './companyUpdateWebsite';
import type { CompanyUpdateDescription } from './companyUpdateDescription';
import type { CompanyUpdatePrimaryColor } from './companyUpdatePrimaryColor';
import type { CompanyUpdateSecondaryColor } from './companyUpdateSecondaryColor';
import type { CompanyUpdateAccentColor } from './companyUpdateAccentColor';

export interface CompanyUpdate {
  name?: CompanyUpdateName;
  industry?: CompanyUpdateIndustry;
  size?: CompanyUpdateSize;
  website?: CompanyUpdateWebsite;
  description?: CompanyUpdateDescription;
  primary_color?: CompanyUpdatePrimaryColor;
  secondary_color?: CompanyUpdateSecondaryColor;
  accent_color?: CompanyUpdateAccentColor;
}